Refactor parser into multiple modules
